website: bucket-level tests (indexdoc, errordoc)
This adds tests for the new S3 Website code under development at on the
wip-static-website-robbat2-master branch (see the robbat2 fork).

The ErrorDoc tests with a valid uploaded ErrorDoc show that my existing
code fails as expected :-). All other IndexDoc cases pass already.

It would be nice to refactor some of this set of tests and extract
common code, because the only variation in some of them is the ACLs on a
single file.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>